Calculates projected inventories and coverages

Calculates projected inventories and coverages

light_proj_inv(dataset, DFU, Period, Demand, Opening, Supply)

Arguments

dataset

a dataframe with the demand and supply features for an item per period

DFU

name of an item, a SKU, or a node like an item x location

Period

a period of time monthly or weekly buckets for example

Demand

the quantity of an item planned to be consumed in units for a given period

Opening

the opening inventories of an item in units at the beginning of the horizon

Supply

the quantity of an item planned to be supplied in units for a given period

Value

a dataframe with the calculated projected inventories and coverages and the related analysis

Examples

light_proj_inv(dataset = blueprint_light, DFU, Period, Demand, Opening, Supply)
#> Joining, by = c("DFU", "Period")
#> # A tibble: 520 × 7
#> # Groups:   DFU [10]
#>    DFU         Period     Demand Opening Calculated.Coverage.in…¹ Proje…² Supply
#>    <chr>       <date>      <dbl>   <dbl>                    <dbl>   <dbl>  <dbl>
#>  1 Item 000001 2022-07-03    364    6570                       17    6206      0
#>  2 Item 000001 2022-07-10    364       0                       16    5842      0
#>  3 Item 000001 2022-07-17    364       0                       15    5478      0
#>  4 Item 000001 2022-07-24    260       0                       14    5218      0
#>  5 Item 000001 2022-07-31    736       0                       13    4482      0
#>  6 Item 000001 2022-08-07    859       0                       12    3623      0
#>  7 Item 000001 2022-08-14    859       0                       11    2764      0
#>  8 Item 000001 2022-08-21    859       0                       10    1905      0
#>  9 Item 000001 2022-08-28    273       0                        9    1632      0
#> 10 Item 000001 2022-09-04    349       0                        8    1283      0
#> # … with 510 more rows, and abbreviated variable names
#> #   ¹​Calculated.Coverage.in.Periods, ²​Projected.Inventories.Qty