Calculates projected inventories and coverages and perform an analysis vs stocks targets

Calculates projected inventories and coverages and perform an analysis vs stocks targets

proj_inv(
  dataset,
  DFU,
  Period,
  Demand,
  Opening.Inventories,
  Supply.Plan,
  Min.Stocks.Coverage,
  Max.Stocks.Coverage
)

Arguments

dataset

a dataframe with the demand and supply features for an item per period

DFU

name of an item, a SKU, or a node like an item x location

Period

a period of time monthly or weekly buckets for example

Demand

the quantity of an item planned to be consumed in units for a given period

Opening.Inventories

the opening inventories of an item in units at the beginning of the horizon

Supply.Plan

the quantity of an item planned to be supplied in units for a given period

Min.Stocks.Coverage

minimum stocks target of an item expressed in periods

Max.Stocks.Coverage

maximum stocks target of an item expressed in periods

Value

a dataframe with the calculated projected inventories and coverages and the related analysis

Examples

proj_inv(dataset = blueprint, DFU = DFU, Period = Period, Demand = Demand, Opening.Inventories = Opening.Inventories, Supply.Plan = Supply.Plan, Min.Stocks.Coverage = Min.Stocks.Coverage, Max.Stocks.Coverage = Max.Stocks.Coverage)
#> Joining, by = c("DFU", "Period")
#> Joining, by = c("DFU", "Period")
#> # A tibble: 520 × 14
#> # Groups:   DFU [10]
#>    DFU         Period     Demand Openi…¹ Calcu…² Proje…³ Suppl…⁴ Min.S…⁵ Max.S…⁶
#>    <chr>       <date>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 Item 000001 2022-07-03    364    6570      17    6206       0       4       8
#>  2 Item 000001 2022-07-10    364       0      16    5842       0       4       8
#>  3 Item 000001 2022-07-17    364       0      15    5478       0       4       8
#>  4 Item 000001 2022-07-24    260       0      14    5218       0       4       8
#>  5 Item 000001 2022-07-31    736       0      13    4482       0       4       8
#>  6 Item 000001 2022-08-07    859       0      12    3623       0       4       8
#>  7 Item 000001 2022-08-14    859       0      11    2764       0       4       8
#>  8 Item 000001 2022-08-21    859       0      10    1905       0       4       8
#>  9 Item 000001 2022-08-28    273       0       9    1632       0       4       8
#> 10 Item 000001 2022-09-04    349       0       8    1283       0       4       8
#> # … with 510 more rows, 5 more variables: Safety.Stocks <dbl>,
#> #   Maximum.Stocks <dbl>, PI.Index <chr>, Ratio.PI.vs.min <dbl>,
#> #   Ratio.PI.vs.Max <dbl>, and abbreviated variable names ¹​Opening.Inventories,
#> #   ²​Calculated.Coverage.in.Periods, ³​Projected.Inventories.Qty, ⁴​Supply.Plan,
#> #   ⁵​Min.Stocks.Coverage, ⁶​Max.Stocks.Coverage